Implemented SignUp feature
The LoginSignUpModal stores the email and password typed into the
forms by the users and sends it to Node upon SignUp submission.
Node calls Firebase and creates an entry for the user with a unique
user id. Login not yet implemented. Node does not check for existing
accounts. Data is transferred between LoginSignUpModal (child) and
CustomNavbar (parent) through callback prop.