chat: Converted profile section into a sidebar with two subsections
Previous implementation had our "profile" component take up the entire right side portion of the website, essentially acting as a sidebar.

Rather than having our chat window be a child of the "profile" section, I create a sidebar <div> with "profile" and "ChatWindow" components be the children.

This should prevent issues when modifying the CSS and overall layout of the sidebar section.