Added display names (i.e., usernames)
For each user authenticated by firebase, their username is also
stored in the displayName field. When signed up or logged in, this
field is passed to the Scramboard component from the LoginSignupModal,
which is then sent to the necessary child components (CustomNavbar,
ChatWindow, ...). So, the CustomNavbar displays the username when
logged in and the ChatWindow displays the user's username before their
chat messages.

TODO: logged-out users should not be able to change the Board.