Bug fix: App crashes after login
The issue was that the username instance in Scramboard component
is null when the user is not logged in. The code depends on this
fact. The initial value was changed by Saman to "Anon", which
broke other parts of the program. This change is needed for
his user Avatar implementation. So, I made the user avatar
render iff user is logged in.