Feature: Added board snapshot look-up
Created a "history" field in firebase. In node,
"history" is initialized to its default values
if needed. By default, "history" saves the number
of edits (or clicks) made thus far to the board.
It saves the row and column number and the new color
associated with a click.

Whenever the client calls the board POST API to write
to the board in Firebase, the board POST API now also
edits the history field (on top of the pixel field).

I created a new snapshot POST API. This API
receives a clickNumber and returns a board snapshot
that represents what the board looked like after the
clickNumber of clicks. I construct this list using the
information in the history field. Therefore, this new
feature has very little overhead on coloring pixels,
but traveling back in time requires O(clickNum) to
construct the board.

Added a stylesheet for the ColorPicker component.

Added a form under the ColorPicker component that
accepts any clickNumber. Note that this form is in
the Scramboard component. When a new clickNumber
is submitted, the Scramboard component sends this
information to the Board component. The board
component then displays a snapshot of the board
after the clickNumber amount of clicks.

Lastly, I added an Alert message when the user attempts
to edit a snapshot of the board. Thus, they are prompted
to go back to the present version before continuing.

TODO: 1. handle invalid clickNumbers (I.e. those not in the
         board history)
      2. Change form into a snapshot list (i.e. instead of
         a form accepting a clickNumber, create a list of all
         the snapshots of the board that the user clicks on).

Known issue: When the clickNumber prop changes, the
Board component is not re-rendered. The board waits for the
interval call to setBoardFromDB() to re-render.