Bug fix: <= 5 sec delay between snapshot switching
Originally, when the clickNumber prop of the Board component
was changed, the component's board was not updated because the
setBoardFromDB function gets called every 5 seconds. Now, the
Board component has a new instance called this.renderedClickNum,
which holds the current clickNumber rendered by the board (i.e.
the snapshot presented on the board). If this.renderedClickNum
defers from this.prop.clickNumber during a render call, then the
setBoardFromDB is called again (instead of waiting 5 seconds for
it to be called automatically).