Feature: Added scrollable list for snapshots
Before, the snapshots were viewed using forms (i.e., text boxes
that accepted the snapshot number that the user wanted to
travel back to). However, the user can now choose the snapshot
to travel back to in a list of scrollable snapshots on the left
side of the main page.

The same functionalities are still intact for snapshots. If a user
attempts to edit snapshots, an alert appears, which informs them
to go back to the live board.

I created a new GET API that returns the number of snapshots to the
front-end application.