chat: Improved chatbox look and added emoji support
Improved the look of the chatbox so that it looks cleaner than what it looked like before. Added a border around it to "designate" it as its own component on the site. Usernames will have a random color associated with them; though, a current bug is that every time the site is updated the color of the usernames changes.

Since chat messaging systems nowadays allows users to insert emojis into their messages, I went ahead and implemented "emoji support." Already, React can process the emojis into an image in the text, so when I mean by "emoji support," it's more like allowing users to select emojis from a menu and inserting them into their messages.

This was done using a React component made by Cesar William Alvarenga (aka cesarwbr) called "<InputEmoji/>". Essentially, it's a chat entry React component that is connected to an emoji explorer, allowing users to look up emojis and inserting them into their chat message. The source of this component can be found in the following link: https://www.npmjs.com/package/react-input-emoji

As a small change to the project structure, I placed a "Messages.css" file in /client/src/MainPage/Messages directory to keep the CSS stylesheets for the chatbox components separate from the overall Scramboard components.