UI Improvements: snapshots, status bar, and board
Improved UI of snapshot lists and the right status bar by adding
borders and shadows. Added hover effect to the snapshot buttons.
Added shadow to the Board component.

The AlertMessage component was not passing the className prop
to the Alert component, so defining a class for AlertMessage
in another component that uses alerts did not work. I simply
passed the className prop of AlertMessage to its child classes.

Furthermore, when logging in, instead of shifting everything down,
the login/signup success/fail alerts will be on top of the board.