Implemented a bot to generate pixel art from image
This bot is implemented for the demo, so we can preview an already
made pixel art. This bot is written in Python and uses the Pillow
and NumPy modules. It takes as input an image of resolution 50x50
and converts it to a 2d array of rgb values. It outputs this array
in a format that can be read by JavaScript.

I created a bot.js module in Node that populates the board with
the pixels from the bot. The line bot.run(db) is commented in
the index.js file. Uncomment this line to run the bot. Do not run
the bot every time the server is launched since any previous
history is erased.