Automated the bot for generating pixelated images
The pixel_array_generator.py script now reads the image.png file,
converts it to a 50x50 image, and outputs its pixels to the
PixelArrayData.txt file. This file is then read by the bot.js
script, which is invoked from Node's index.js. Thereby, the
data is automatically passed from the python script to Node without
the need to copy and paste the array to the source code (which was
the case before).