Merged with origin/heatmap + click timestamps
Changed firebase timestamps for clicks and altered the timestamps
generated by the bot.

Changed history/clicks/date to history/clicks/time in firebase.
Changed time entry to epoch time in the initialization Node script.
The bot now prompts for the image name to be pixelated onto the
board. The clicks generated by the bot are now uploaded to the DB
in such a way that the user thinks there were 100 clicks per day
for the span of the entire 2500 clicks (50x50 board) to be
generated in the present day. This will allow for a more intuitive
heatmap demonstration.